Moved the metric class
