--- src/main/java/io/github/thebusybiscuit/slimefun4/core/services/metrics/package-info.java
@@ -1,4 +1,0 @@
-/**
- * This package contains everything related to bStats Metrics
- */
-package io.github.thebusybiscuit.slimefun4.core.services.metrics;
